abstracted out api call from component
