+++ package.json
@@ -8,7 +8,6 @@
     "cypress:open": "cypress open",
     "cypress:run": "cypress run",
     "dev": "webpack-dev-server --open --mode=development",
-    "precommit": "lint-staged",
     "prepush": "jest",
     "storybook": "start-storybook -p 6009",
     "build-storybook": "build-storybook",
