+++ src/index.jsx
@@ -6,17 +6,16 @@ import BounceRulesContainer from "./components/BounceRulesContainer";
 import BounceRuleDetails from "./components/BounceRuleDetails";
 import "./index.scss";
 
-const App = () => <div />;
-
-ReactDOM.render(
+const App = () => (
   <BrowserRouter>
     <div>
       <Route exact path="/" component={Login} />
       <Route exact path="/bounce_rules" component={BounceRulesContainer} />
       <Route path="/bounce_rules/:bounceRuleId" component={BounceRuleDetails} />
     </div>
-  </BrowserRouter>,
-  document.getElementById("root")
+  </BrowserRouter>
 );
+
+ReactDOM.render(<App />, document.getElementById("root"));
 
 export default App;
